Remove leftovers
